+++ memory/mutation_boundaries_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/faustbrian/golib/pkg/audit"
+	"github.com/faustbrian/go-audit"
 )
 
 func TestStoreValidationFailuresLeaveUsableState(t *testing.T) {