+++ record_contract_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/faustbrian/golib/pkg/audit"
+	"github.com/faustbrian/go-audit"
 )
 
 func TestRecordPreservesCompleteContextAndIdentitySemantics(t *testing.T) {